requirements.txt
discord.py-self
urlextract